+++ src/pages/account/profile/addresses/new.tsx
@@ -10,7 +10,7 @@ import type { GetServerSideProps } from "next";
 import React from "react";
 
 import { Container } from "@src/components/containers/container";
-import { UserAddressForm } from "@src/components/forms/profile/update-user-address-form";
+import { UserAddressForm } from "@src/components/forms/profile/user-address-form";
 import { Layout } from "@src/components/page-layout";
 import { configVariables } from "@src/configs/config-variables";
 import { logger } from "@src/configs/logger";
